requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.0.0
anthropic>=0.40.0
requests>=2.31.0
# content fetching
trafilatura>=1.12.0
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0
# transcription (requires ffmpeg installed on system)
faster-whisper>=1.0.0
# documents
pdfplumber>=0.11.0
